requirements.txt
beautifulsoup4==4.12.2
datasets==2.14.3
huggingface-hub==0.16.4
ipython==8.14.0
ipywidgets==8.1.1
jedi==0.19.0
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.24.0
langchain==0.0.310
langcodes==3.3.0
langsmith==0.0.43
llama-hub==0.0.47post1
llama-index==0.9.6.post1
loguru==0.7.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
numpy==1.24.4
openai==1.3.5
pandas==2.0.3
protobuf==4.23.4
pyarrow==12.0.1
python-dotenv==1.0.0
rank-bm25==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
sentence-transformers==2.2.2
streamlit==1.28.2
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.1
weaviate-client==3.25.3 
wikipedia
